requirements.txt
pdftotext==2.1.4
Django>=2.1,<2.2
djangorestframework==3.11.0
django-environ==0.4.5
wtforms